+++ oap-server/analyzer/meter-analyzer/src/main/java/org/apache/skywalking/oap/meter/analyzer/dsl/FilterExpression.java
@@ -51,4 +51,8 @@ public class FilterExpression {
         }
         return sampleFamilies;
     }
+
+    public SampleFamily filter(SampleFamily sampleFamily) {
+        return sampleFamily.filter(filterClosure);
+    }
 }